Finish flight ingress refactor plan
